+++ include/c74_min_operator_matrix.h
@@ -167,13 +167,13 @@ namespace min {
 	template<class min_class_type>
 	void* max_jit_mop_new(max::t_symbol* s, long argc, max::t_atom* argv) {
 		atom_reference		args(argc, argv);
-        long				attrstart	= attr_args_offset(args.size(), args.begin());
+        long				attrstart	= attr_args_offset((short)args.size(), args.begin());
 		auto				cppname		= (max::t_symbol*)this_class->c_menufun;
 		max_jit_wrapper*	self		= (max_jit_wrapper*)max::max_jit_object_alloc(this_class, cppname);
 		void*				o			= max::jit_object_new(cppname, s);
 		
 		max_jit_mop_setup_simple(self, o, args.size(), args.begin());
-		max_jit_attr_args(self, args.size(), args.begin());
+		max_jit_attr_args(self, (short)args.size(), args.begin());
         
         auto	job = (minwrap<min_class_type>*)o;
         job->min_object.try_call("maxob_setup", atoms(args.begin(), args.begin()+attrstart));
